refactor: move queryPersonaByMnemonic
