+++ packages/mask/background/services/identity/index.ts
@@ -12,6 +12,7 @@ export {
     loginExistPersonaByPrivateKey,
     mobile_queryPersonaByPrivateKey,
     renamePersona,
+    queryPersonaByMnemonic,
 } from './persona/update'
 export { signWithPersona, type SignRequest, type SignRequestResult, generateSignResult } from './persona/sign'
 export { exportPersonaMnemonicWords, exportPersonaPrivateKey } from './persona/backup'
@@ -30,3 +31,5 @@ export {
 export { createNewRelation } from './relation/create'
 export { mobile_queryRelationsRecordFromIndexedDB } from './relation/query'
 export { updateRelation } from './relation/update'
+
+export { validateMnemonic } from './persona/utils'
